+++ contracts/emission_setters/Constant.sol
@@ -3,7 +3,7 @@
 pragma solidity 0.8.13;
 
 contract Constant {
-    function getRate() external pure returns (uint256 rate) {
+    function getRate() external pure returns (uint rate) {
         rate = 1e17;
     }
 }
